requirements.txt
Flask==2.3.2
tensorflow==2.10.0
numpy==1.24.2
Pillow==9.4.0
Werkzeug==2.3.7
opencv-python==4.7.0.72
scikit-learn==1.2.2